:sparkles: responde a message with the zerotier id
